click typer/requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
python-dotenv
tenacity
